backend/requirements.txt
flask==3.1.1
flask-cors==6.0.1
pymongo==4.13.0
python-dotenv==1.1.0
parameterized==0.9.0